+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/components/estimators/AbstractTripRouterEstimator.java
@@ -4,12 +4,14 @@ import java.util.List;
 
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.PlanElement;
+import org.matsim.core.config.Config;
 import org.matsim.core.router.TripRouter;
 import org.matsim.facilities.ActivityFacilities;
 import org.matsim.facilities.FacilitiesUtils;
 import org.matsim.facilities.Facility;
 
 import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceTrip;
+import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceUtils;
 import ch.ethz.matsim.discrete_mode_choice.model.trip_based.TripEstimator;
 import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.DefaultRoutedTripCandidate;
 import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.TripCandidate;
@@ -23,10 +25,12 @@ import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.TripCandi
  * @author sebhoerl
  */
 public abstract class AbstractTripRouterEstimator implements TripEstimator {
+	private final Config config;
 	private final TripRouter tripRouter;
 	private final ActivityFacilities facilities;
 
-	public AbstractTripRouterEstimator(TripRouter tripRouter, ActivityFacilities facilities) {
+	public AbstractTripRouterEstimator(Config config, TripRouter tripRouter, ActivityFacilities facilities) {
+		this.config = config;
 		this.tripRouter = tripRouter;
 		this.facilities = facilities;
 	}
@@ -35,7 +39,7 @@ public abstract class AbstractTripRouterEstimator implements TripEstimator {
 	public final TripCandidate estimateTrip(Person person, String mode, DiscreteModeChoiceTrip trip,
 			List<TripCandidate> previousTrips) {
 		// I) Find the correct origin and destination facilities
-		
+
 		Facility originFacility = FacilitiesUtils.toFacility(trip.getOriginActivity(), facilities);
 		Facility destinationFacility = FacilitiesUtils.toFacility(trip.getDestinationActivity(), facilities);
 
@@ -63,6 +67,8 @@ public abstract class AbstractTripRouterEstimator implements TripEstimator {
 	protected TripCandidate estimateTripCandidate(Person person, String mode, DiscreteModeChoiceTrip trip,
 			List<TripCandidate> previousTrips, List<? extends PlanElement> routedTrip) {
 		double utility = estimateTrip(person, mode, trip, previousTrips, routedTrip);
-		return new DefaultRoutedTripCandidate(utility, mode, routedTrip);
+		double duration = DiscreteModeChoiceUtils.advanceTime(routedTrip, trip.getDepartureTime(), config);
+
+		return new DefaultRoutedTripCandidate(utility, mode, routedTrip, duration);
 	}
 }