+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/components/estimators/UniformTourEstimator.java
@@ -1,15 +1,18 @@
 package ch.ethz.matsim.discrete_mode_choice.components.estimators;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.matsim.api.core.v01.population.Person;
+import org.matsim.core.config.Config;
 
 import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceTrip;
+import ch.ethz.matsim.discrete_mode_choice.model.DiscreteModeChoiceUtils;
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.DefaultTourCandidate;
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourCandidate;
 import ch.ethz.matsim.discrete_mode_choice.model.tour_based.TourEstimator;
 import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.DefaultTripCandidate;
+import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.TripCandidate;
 
 /**
  * This estimator simply return a zero utility for every tour candidate that it
@@ -18,10 +21,24 @@ import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.DefaultTr
  * @author sebhoerl
  */
 public class UniformTourEstimator implements TourEstimator {
+	private final Config config;
+
+	public UniformTourEstimator(Config config) {
+		this.config = config;
+	}
+
 	@Override
 	public TourCandidate estimateTour(Person person, List<String> modes, List<DiscreteModeChoiceTrip> trips,
 			List<TourCandidate> previousTours) {
-		return new DefaultTourCandidate(1.0,
-				modes.stream().map(mode -> new DefaultTripCandidate(1.0, mode)).collect(Collectors.toList()));
+		List<TripCandidate> tripCandidates = new ArrayList<>(modes.size());
+
+		for (int index = 0; index < modes.size(); index++) {
+			DiscreteModeChoiceTrip trip = trips.get(index);
+			double duration = DiscreteModeChoiceUtils.advanceTime(trip.getInitialElements(), trip.getDepartureTime(),
+					config);
+			tripCandidates.add(new DefaultTripCandidate(1.0, modes.get(index), duration));
+		}
+
+		return new DefaultTourCandidate(1.0, tripCandidates);
 	}
 }