+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/trip_based/candidates/DefaultRoutedTripCandidate.java
@@ -12,8 +12,9 @@ import org.matsim.api.core.v01.population.PlanElement;
 public class DefaultRoutedTripCandidate extends DefaultTripCandidate implements RoutedTripCandidate {
 	private final List<? extends PlanElement> routedPlanElements;
 
-	public DefaultRoutedTripCandidate(double utility, String mode, List<? extends PlanElement> routedPlanElements) {
-		super(utility, mode);
+	public DefaultRoutedTripCandidate(double utility, String mode, List<? extends PlanElement> routedPlanElements,
+			double duration) {
+		super(utility, mode, duration);
 		this.routedPlanElements = routedPlanElements;
 	}