+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/trip_based/candidates/TripCandidate.java
@@ -9,4 +9,6 @@ import ch.ethz.matsim.discrete_mode_choice.model.utilities.UtilityCandidate;
  */
 public interface TripCandidate extends UtilityCandidate {
 	String getMode();
+	
+	double getDuration();
 }